requirements.txt
langchain==0.2.16
langchain-openai==0.1.23
openai==1.40.0
pandas==2.1.4
openpyxl==3.1.5
streamlit==1.37.0
python-dotenv==1.0.1
numpy==1.26.4
apscheduler==3.10.4
